sstate-publish: only rsync our sstate, not our base sstate any more
This had to be done when we were using hardlink-based CoW on sstates,
but now that we're using SSTATE_MIRRORS instead there's no need to keep
it (even if the given sstate is also touched by old hardlink-based scripts).

Since we want to use --safe-links it is safer to stop sync'ing it now, as
defensive future-proofing.